requirements.txt
streamlit
requests
websockets
pandas
python-dotenv
streamlit-autorefresh
web3